appointment/requirements.txt
Django==3.2.23
django-model-utils==4.2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
PyJWT==1.6.4
django-cors-headers==3.13.0
django-filter==2.4.0
psycopg2==2.8.4
sentry-sdk==1.14.0
